Uses integer values for x and y in the GetFeatureInfo request. Do not try to search snapshots version on some repositories that are not needed.

+++ web-services/services/src/main/java/org/constellation/provider/postgrid/PostGridLayerDetails.java
@@ -150,8 +150,8 @@ class PostGridLayerDetails implements LayerDetails {
         }
 
         // Pixel coordinates in the request.
-        final double pixelUpX     = gfi.getX();
-        final double pixelUpY     = gfi.getY();
+        final int pixelUpX        = gfi.getX();
+        final int pixelUpY        = gfi.getY();
         final double widthEnv     = objEnv.getSpan(0);
         final double heightEnv    = objEnv.getSpan(1);
         final double resX         =      widthEnv  / width;
